a little bit in biodiversity
Nevertheless, Hubbell's neutral theory is contentious. 
The idea that each species competes with all others to the same degree as intraspecies competition strains credibility. 
Surely the differences between species matters! 
Of course there are differences between species; even the staunchest neutralist would agree. 
But slight perturbations from Hubbell's theory do not significantly alter its results. 
What's more, while everyone concedes that there are differences between species, some argue that these differences do not matter. 
In some sense, they claim, the species are equivalent and behave neutrally, which is why Hubbell's theory seems to work so well in such disparate ecologies. 
The examples presented in Hubbell's seminal book are compelling, and their may be some truth to these claims. 
The other side of the debate insists that species differences are the cause of observed abundance distributions. 
In particular, the environment can be divided into various ecological niches, and it is how these niches are uniquely used by their occupying species that determines the biodiversity of the ecosystem. 
This is broadly known as niche theory. 
%It itself has a contentious past mired by confusions. I will do my best to provide a summary here, to contextualize my research.